selftest/buildhistory: Move test cases to common location
Having the test cases split in different places makes no sense. Move them
all to the place you'd expect to find them. Drop the base class as it
is no longer needed.

(From OE-Core rev: 764780385a29bc7090d74e9b34ce8fa7c1c1a57f)
